+++ src/App.css
@@ -1,47 +1,98 @@
+:root{
+    --headerBackgroundColor:#222;
+    --colorWhite:white;
+}
+
 .navbar{
-    background-color: #222;
+    background-color: var(--headerBackgroundColor);
     display: flex;
     align-items: center;
     justify-content: space-between;
     padding: 20px 6%;
-    color: white;
+    color: var(--colorWhite);
+}
+
+.navbar > a{
+    height: 64px;
 }
 
 .navbar__logo{
-    width: 64px;
-    height: 64px;
+    width: 100%;
+    height: 100%;
 }
 
 .navbar__links{
     display: flex;
     list-style: none;
     gap: 20px;
+    text-align: center;
+    
+    & a {
+        color: var(--colorWhite);
+        text-decoration: none;
+        font-weight: bold;
+        cursor:pointer
+    }
 }
 
-.menuToggle{
-    width: 35px;
+.navbar__links a:hover{
+    color: yellow;
+}
+
+.navbar__links--dropdown{
     display: none;
+    height: 0;
+    flex-direction: column;
+    gap: 0;
+    position: absolute;
+    right: 0;
+    top: 104px;
+    background-color: var(--headerBackgroundColor);
+    overflow: hidden;
+    transition: .5s;
+}
+
+.navbar__links--dropdown.show{
+    height: 208px;
+}
+
+.navbar__links--dropdown a{
+    padding:15px 20px;
+    cursor:pointer;
+    transition: .5s;
+    
+    &:hover{
+        color: var(--headerBackgroundColor);
+        background-color: var(--colorWhite);
+    }
+}
+ 
+.menuToggle{
+    width: 25px;
+    display: none;
+    cursor: pointer;
+    transition: .5s;
+    flex-direction: column;
+    align-items: center;
+    gap: 5px;
     cursor: pointer;
 }
 
 .clicked .menuToggle__line:nth-of-type(1){
-    transform: translate(0,4px) rotate(44deg);
+    transform:translate(0, 8px) rotate(44deg);
 }
 
 .clicked .menuToggle__line:nth-of-type(3){
-    transform: rotate(-44deg);
+    transform:translate(0, -8px) rotate(-44deg);
 }
 
 .clicked .menuToggle__line:nth-of-type(2){
-    display: none;
+    width: 0;
 }
 
 .menuToggle__line{
     width: 100%;
-    height: 4px;
-    background-color: white;
-
-    &:nth-of-type(2){
-        margin: 5px 0;
-    }
+    height: 3px;
+    background-color: var(--colorWhite);
+    transition: .6s;
 }
