+++ src/Breakpoints.css
@@ -1,10 +1,60 @@
 @media screen and (max-width:600px){
+    .navbar{
+        padding: 15px 6%;
+    }
+
+    .navbar > a{
+        height: 56px;
+    }
+
     .navbar__links{
         display: none;
     }
 
+    .navbar__links--dropdown{
+        top: 86px;
+        display: flex;
+    }
+
     .menuToggle{
-        display: block;
+        display: flex;
         transition: .5s;
     }
+}
+
+@media screen and (max-width:385px) {
+    .navbar{
+        padding: 10px 6%;
+    }
+
+    .navbar > a {
+        height: 40px;
+    }
+
+    .menuToggle{
+        width: 22px;
+        gap: 4px;
+    }
+
+    .navbar__links--dropdown{
+        top: 60px;
+    }
+
+    .navbar__links--dropdown.show{
+        height: 176px;
+    }
+
+    .navbar__links--dropdown a{
+        padding: 12px 15px;
+        font-size: 15px;
+    }
+
+    .clicked .menuToggle__line:nth-of-type(1){
+        transform:translate(0, 6px) rotate(44deg);
+    }
+
+    .clicked .menuToggle__line:nth-of-type(3){
+        transform:translate(0, -7.5px) rotate(-44deg);
+    }
+    
 }
